+++ packages/dendron-next-server/pages/vscode/note-preview.tsx
@@ -230,7 +230,6 @@ function Note({ engine, ide, ws, port }: DendronProps & WorkspaceProps) {
     ws,
     port,
   });
-  debugger;
   return (
     <AntLayout>
       <Head>